Исправил clientModel поле id,нэйминг переменных в clientModel (chat_id, pet_id)

+++ src/main/java/pro/sky/telegrambot/repository/ClientRepository.java
@@ -3,5 +3,5 @@ package pro.sky.telegrambot.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pro.sky.telegrambot.model.ClientModel;
 
-public interface ClientRepository extends JpaRepository<ClientModel, Long> {
+public interface ClientRepository extends JpaRepository<ClientModel, Integer> {
 }
